Extract MVT encoder into its own module.
`vector.py` is now orchestration only; the numba command-stream encoder +
protobuf framing live in `render/mvt.py` behind a small functional API
(`quantize_rings`, `encode_mvt_layer`, `encode_mvt_tile`). The new module's
docstring records why we kept the hand-rolled encoder over mapbox-vector-tile,
vt2pbf, python-vtzero, and rio-tiler-mvt — none of them offered a numpy batch
geometry path with typed numeric properties at our worst-case feature count.

Tile bytes are byte-identical to the prior implementation.
